# pkg/router/users_router.go
package router

import (
	"github.com/6156-DonaldDuck/users/pkg/config"
	"github.com/6156-DonaldDuck/users/pkg/model"
	"github.com/6156-DonaldDuck/users/pkg/service"
	"github.com/gin-gonic/gin"
	"github.com/gin-contrib/cors"
	log "github.com/sirupsen/logrus"
	"net/http"
	"strconv"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	docs "github.com/6156-DonaldDuck/users/docs"
)


func InitRouter() {
	r := gin.Default()
	r.Use(cors.Default()) // default allows all origin
	docs.SwaggerInfo.BasePath = config.Configuration.Mysql.Host

	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	// users
	r.GET( "/api/v1/users", ListUsers)
	r.GET( "/api/v1/users/:userId", GetUserById)
	r.POST( "/api/v1/users", CreateUser)
	r.PUT( "/api/v1/users/:userId", UpdateUserById)
	r.DELETE( "/api/v1/users/:userId", DeleteUserById)
	// addresses
	r.GET( "/api/v1/addresses", ListAddresses)
	r.GET( "/api/v1/addresses/:addressId", GetAddressById)
	r.POST( "/api/v1/addresses", CreateAddress)
	r.PUT( "/api/v1/addresses/:addressId", UpdateAddressById)
	r.DELETE( "/api/v1/addresses/:addressId", DeleteAddressById)
	// // user + address
	r.GET( "/api/v1/users/:userId/address", GetAddressByUserId)
	// r.POST( "/api/v1/users/:userId/address", SetAddressByUserId)
	// r.GET( "/api/v1/addresses/:addressId/users", ListUsersByAddressId)
	// not sure what this means
	// r.POST( "/api/v1/addresses/:id/users", ) 
	r.Run(":" + config.Configuration.Port)
}

// @BasePath /api/v1

// @Summary List All Users
// @Schemes
// @Description List all users
// @Tags Users
// @Accept json
// @Produce json
// @Success 200 {json} users
// @Failure 500 internal server error
// @Router /users [get]
func ListUsers(c *gin.Context) {
	users, err := service.ListUsers()
	if err != nil {
		c.JSON(http.StatusInternalServerError, "internal server error")
	} else {
		c.JSON(http.StatusOK, users)
	}
}

// @Summary Get User By User Id
// @Schemes
// @Description Get user by user id
// @Tags Users
// @Accept json
// @Produce json
// @Param ID path int true "the id of a specfic user"
// @Success 200 {json} user
// @Failure 400 invalid user id
// @Router /users/{userId} [get]
func GetUserById(c *gin.Context) {
	idStr := c.Param("userId")
	userId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.GetUserById] failed to parse user id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid user id")
		return
	}
	user, err := service.GetUserById(uint(userId))
	if err != nil {
		c.JSON(http.StatusInternalServerError, err.Error())
	} else {
		c.JSON(http.StatusOK, user)
	}
}

// @Summary Delete User By User Id
// @Schemes
// @Description Delete user by user id
// @Tags Users
// @Accept json
// @Produce json
// @Param ID query int true "the id of a specfic user"
// @Success 200 {json} delete successfully
// @Failure 400 invalid user id
// @Router /users/ [delete]
func DeleteUserById(c *gin.Context){
	idStr := c.Param("userId")
	userId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.DeleteUserById] failed to parse user id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid user id")
		return
	}
	err = service.DeleteUserById(uint(userId))
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, "Successfully delete user with id "+ idStr)
	}
}

// @Summary Create User
// @Schemes
// @Description Create User
// @Tags Users
// @Accept json
// @Produce json
// @Param first_name query string false "First Name"
// @Param last_name query string false "Last Name"
// @Param phone_number query string false "Phone Number"
// @Param email query string false "Email"
// @Param address_id query int false "Address ID"
// @Success 200 {json} user id
// @Failure 400 invalid user id
// @Router /users/ [post]
func CreateUser(c *gin.Context){
	user := model.User{}
	if err := c.ShouldBind(&user); err != nil{
		c.JSON(http.StatusBadRequest, err)
	}
	userId, err := service.CreateUser(user)
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, userId)
	}
}

// @Summary Update User By User Id
// @Schemes
// @Description Update user by user id
// @Tags Users
// @Accept json
// @Produce json
// @Param ID path int true "the id of a specfic user"
// @Success 200 {json} update successfully
// @Failure 400 invalid user id
// @Router /users/{userId} [put]
func UpdateUserById(c *gin.Context){
	idStr := c.Param("userId")
	userId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.UpdateUserById] failed to parse user id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid user id")
		return
	}
	updateInfo := model.User{}
	if err := c.ShouldBind(&updateInfo); err != nil{
		c.JSON(http.StatusBadRequest, err)
	}
	updateInfo.ID = uint(userId)
	err = service.UpdateUser(updateInfo)
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, "update successfully")
	}
}

// @BasePath /api/v1

// @Summary List All Addresses
// @Schemes
// @Description List all addresses
// @Tags Addresses
// @Accept json
// @Produce json
// @Success 200 {json} addresses
// @Failure 500 internal server error
// @Router /addresses [get]
func ListAddresses(c *gin.Context) {
	users, err := service.ListAddresses()
	if err != nil {
		c.JSON(http.StatusInternalServerError, "internal server error")
	} else {
		c.JSON(http.StatusOK, users)
	}
}

// @Summary Get Address By Address Id
// @Schemes
// @Description Get addresses by addresses id
// @Tags Addresses
// @Accept json
// @Produce json
// @Param ID path int true "the id of a specfic addresses"
// @Success 200 {json} addresses
// @Failure 400 invalid addresses id
// @Router /addresses/{addresseId} [get]
func GetAddressById(c *gin.Context) {
	idStr := c.Param("addressId")
	addressId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.GetAddressById] failed to parse address id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid address id")
		return
	}
	address, err := service.GetAddressById(uint(addressId))
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, address)
	}
}

// @Summary Create Address
// @Schemes
// @Description Create Address
// @Tags Addresses
// @Accept json
// @Produce json
// @Param street_number query string false "Street Number"
// @Param street_name_1 query string false "Street Name Line 1"
// @Param street_name_2 query string false "Street Name Line 2"
// @Param city query string false "City"
// @Param region query string false "Region"
// @Param country_code query string false "Country Code"
// @Param postal_code query string false "Postal Code"
// @Success 200 {json} address id
// @Failure 400 invalid address id
// @Router /addresses/ [post]
func CreateAddress(c *gin.Context){
	address := model.Address{}
	if err := c.ShouldBind(&address); err != nil{
		c.JSON(http.StatusBadRequest, err)
	}
	addressId, err := service.CreateAddress(address)
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, addressId)
	}
}

// @Summary Update Address By Address Id
// @Schemes
// @Description Update address by address id
// @Tags Addresses
// @Accept json
// @Produce json
// @Param ID path int true "the id of a specfic address"
// @Success 200 {json} update successfully
// @Failure 400 invalid address id
// @Router /addresses/{articleId} [put]
func UpdateAddressById(c *gin.Context){
	idStr := c.Param("addressId")
	addressId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.UpdateAddressById] failed to parse address id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid address id")
		return
	}
	updateInfo := model.Address{}
	if err := c.ShouldBind(&updateInfo); err != nil {
		c.JSON(http.StatusBadRequest, err)
		return
	}
	updateInfo.ID = uint(addressId)
	err = service.UpdateAddressById(updateInfo)
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, updateInfo)
	}
}

// @Summary Delete Address By Address Id
// @Schemes
// @Description Delete address by address id
// @Tags Addresses
// @Accept json
// @Produce json
// @Param ID query int true "the id of a specfic address"
// @Success 200 {json} delete successfully
// @Failure 400 invalid address id
// @Router /addresses/ [delete]
func DeleteAddressById(c *gin.Context){
	idStr := c.Param("addressId")
	addressId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.DeleteAddressById] failed to parse address id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid address id")
		return
	}
	err = service.DeleteAddressById(uint(addressId))
	if err != nil {
		c.Error(err)
	} else {
		c.JSON(http.StatusOK, "Successfully delete address with id "+ idStr)
	}
}

func GetAddressByUserId(c *gin.Context){
	idStr := c.Param("userId")
	userId, err := strconv.Atoi(idStr)
	if err != nil {
		log.Errorf("[router.GetAddressByUserId] failed to parse user id %v, err=%v\n", idStr, err)
		c.JSON(http.StatusBadRequest, "invalid user id")
		return
	}
	address, err := service.GetAddressByUserId(uint(userId))
	if err != nil {
		c.JSON(http.StatusInternalServerError, err.Error())
	} else {
		c.JSON(http.StatusOK, address)
	}
}